requirements.txt
pytorch
transformers
datasets
json
evaluate
pandas
re
faiss-cpu
sentence-transformers
langchain
langchain-community
langchain-huggingface
langchain-text-splitters